Efficiency Baby!

+++ OpenGL_Project/PerlinNoise.h
@@ -26,16 +26,16 @@ static int primes[maxPrimeIndex][3] = {
   { 997169939, 842027887, 423882827 }
 };
 
-static double Noise(int i, int x, int y) {
+static float Noise(int i, int x, int y) {
     int n = x + y * 57;
     n = (n << 13) ^ n;
     int a = primes[i][0], b = primes[i][1], c = primes[i][2];
     int t = (n * (n * n * a + b) + c) & 0x7fffffff;
-    return 1.0 - (double)(t) / 1073741824.0;
+    return 1.0f - (t) / 1073741824.0f;
 }
 
-static double SmoothedNoise(int i, int x, int y) {
-    double corners = (Noise(i, x - 1, y - 1) + Noise(i, x + 1, y - 1) +
+static float SmoothedNoise(int i, int x, int y) {
+    float corners = (Noise(i, x - 1, y - 1) + Noise(i, x + 1, y - 1) +
         Noise(i, x - 1, y + 1) + Noise(i, x + 1, y + 1)) / 16,
         sides = (Noise(i, x - 1, y) + Noise(i, x + 1, y) + Noise(i, x, y - 1) +
             Noise(i, x, y + 1)) / 8,
@@ -43,19 +43,19 @@ static double SmoothedNoise(int i, int x, int y) {
     return corners + sides + center;
 }
 
-static double Interpolate(double a, double b, double x) {  // cosine interpolation
-    double ft = x * 3.1415927,
-        f = (1 - cos(ft)) * 0.5;
+static float Interpolate(float a, float b, float x) {  // cosine interpolation
+    float ft = x * 3.1415927f,
+        f = (1 - cos(ft)) * 0.5f;
     return  a * (1 - f) + b * f;
 }
 
-static double InterpolatedNoise(int i, double x, double y) {
+static float InterpolatedNoise(int i, float x, float y) {
     int integer_X = (int)x;
-    double fractional_X = x - integer_X;
+    float fractional_X = x - integer_X;
     int integer_Y = (int)y;
-    double fractional_Y = y - integer_Y;
+    float fractional_Y = y - integer_Y;
 
-    double v1 = SmoothedNoise(i, integer_X, integer_Y),
+    float v1 = SmoothedNoise(i, integer_X, integer_Y),
         v2 = SmoothedNoise(i, integer_X + 1, integer_Y),
         v3 = SmoothedNoise(i, integer_X, integer_Y + 1),
         v4 = SmoothedNoise(i, integer_X + 1, integer_Y + 1),
@@ -64,17 +64,17 @@ static double InterpolatedNoise(int i, double x, double y) {
     return Interpolate(i1, i2, fractional_Y);
 }
 
-static double ValueNoise_2D(double x, double y) {
+static float ValueNoise_2D(float x, float y) {
     int numX = 512,
         numY = 512,
         numOctaves = 7;
-    double persistence = 0.5;
+    float persistence = 0.5;
 
     int primeIndex = 0;
 
 
-    double total = 0,
-        frequency = pow(2, numOctaves),
+    float total = 0,
+        frequency = (float)pow(2, numOctaves),
         amplitude = 1;
     for (int i = 0; i < numOctaves; ++i) {
         frequency /= 2;
